+++ leaguerule.go
@@ -1,7 +1,7 @@
 package poego
 
 type Rules struct {
-	Id          int16  `json:"id"`
+	Id          int    `json:"id"`
 	Name        string `json:"name"`
 	Description string `json:"description"`
 }
